Remove more unused inject code
Removing a couple assertions that we do not use together with their supporting code.
