+++ server/src/main/java/org/elasticsearch/common/inject/ConstructorInjectorStore.java
@@ -21,6 +21,9 @@ import org.elasticsearch.common.inject.internal.ErrorsException;
 import org.elasticsearch.common.inject.internal.FailableCache;
 import org.elasticsearch.common.inject.spi.InjectionPoint;
 
+import java.lang.reflect.Constructor;
+import java.lang.reflect.InvocationTargetException;
+
 /**
  * Constructor injectors by type.
  *
@@ -65,10 +68,28 @@ class ConstructorInjectorStore {
         );
         MembersInjectorImpl<T> membersInjector = injector.membersInjectorStore.get(type, errors);
 
-        ConstructionProxyFactory<T> factory = new DefaultConstructionProxyFactory<>(injectionPoint);
-
         errors.throwIfNewErrors(numErrorsBefore);
 
-        return new ConstructorInjector<>(factory.create(), constructorParameterInjectors, membersInjector);
+        @SuppressWarnings("unchecked") // the injection point is for a constructor of T
+        final Constructor<T> constructor = (Constructor<T>) injectionPoint.getMember();
+        return new ConstructorInjector<>(new ConstructionProxy<>() {
+            @Override
+            public T newInstance(Object... arguments) throws InvocationTargetException {
+                try {
+                    return constructor.newInstance(arguments);
+                } catch (InstantiationException e) {
+                    throw new AssertionError(e); // shouldn't happen, we know this is a concrete type
+                } catch (IllegalAccessException e) {
+                    // a security manager is blocking us, we're hosed
+                    throw new AssertionError("Wrong access modifiers on " + constructor, e);
+                }
+            }
+
+            @Override
+            public InjectionPoint getInjectionPoint() {
+                return injectionPoint;
+            }
+
+        }, constructorParameterInjectors, membersInjector);
     }
 }
