+++ server/src/main/java/org/elasticsearch/common/inject/Module.java
@@ -24,11 +24,6 @@ package org.elasticsearch.common.inject;
  * <p>
  * Your Module classes can use a more streamlined syntax by extending
  * {@link AbstractModule} rather than implementing this interface directly.
- * <p>
- * In addition to the bindings configured via {@link #configure}, bindings
- * will be created for all methods annotated with {@literal @}{@link Provides}.
- * Use scope and binding annotations on these methods to configure the
- * bindings.
  */
 public interface Module {
 
@@ -36,8 +31,7 @@ public interface Module {
      * Contributes bindings and other configurations for this module to {@code binder}.
      * <p>
      * <strong>Do not invoke this method directly</strong> to install submodules. Instead use
-     * {@link Binder#install(Module)}, which ensures that {@link Provides provider methods} are
-     * discovered.
+     * {@link Binder#install(Module)}.
      */
     void configure(Binder binder);
 }
