--- server/src/main/java/org/elasticsearch/common/inject/ProvidedBy.java
@@ -1,38 +1,0 @@
-/*
- * Copyright (C) 2006 Google Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.elasticsearch.common.inject;
-
-import java.lang.annotation.Retention;
-import java.lang.annotation.Target;
-
-import static java.lang.annotation.ElementType.TYPE;
-import static java.lang.annotation.RetentionPolicy.RUNTIME;
-
-/**
- * A pointer to the default provider type for a type.
- *
- * @author crazybob@google.com (Bob Lee)
- */
-@Retention(RUNTIME)
-@Target(TYPE)
-public @interface ProvidedBy {
-
-    /**
-     * The implementation type.
-     */
-    Class<? extends Provider<?>> value();
-}
