--- server/src/main/java/org/elasticsearch/common/inject/util/Modules.java
@@ -1,32 +1,0 @@
-/*
- * Copyright (C) 2008 Google Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.elasticsearch.common.inject.util;
-
-import org.elasticsearch.common.inject.Module;
-
-/**
- * Static utility methods for creating and working with instances of {@link Module}.
- *
- * @author jessewilson@google.com (Jesse Wilson)
- * @since 2.0
- */
-public final class Modules {
-    private Modules() {}
-
-    public static final Module EMPTY_MODULE = binder -> {};
-
-}
